Seperate files into respective packages
